Fix. Don't justify content in the width limiter
`align-items: center;` doesn't do what I thought it did. However, I noticed
that the cat, the image in the hero, isn't squishing all the way.
Unfortunately, that's the result of fixing the overflow out the left side!
Decided to not try to perfect this behavior yet. I need a better understanding
of flex.